pyproject.toml
[project]
name = "mcp-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.51.0",
    "arxiv>=2.2.0",
    "httpx>=0.28.1",
    "mcp[cli]>=1.9.0",
    "nest-asyncio>=1.6.0",
    "python-dotenv>=1.1.0",
]
